+++ libs/evoral/src/libsmf/smf_private.h
@@ -71,7 +71,7 @@ void smf_init_tempo(smf_t *smf);
 void smf_fini_tempo(smf_t *smf);
 void smf_create_tempo_map_and_compute_seconds(smf_t *smf);
 void maybe_add_to_tempo_map(smf_event_t *event);
-void remove_last_tempo_with_pulses(smf_t *smf, int pulses);
+void remove_last_tempo_with_pulses(smf_t *smf, size_t pulses);
 int smf_event_is_tempo_change_or_time_signature(const smf_event_t *event) WARN_UNUSED_RESULT;
 int smf_event_length_is_valid(const smf_event_t *event) WARN_UNUSED_RESULT;
 int is_status_byte(const unsigned char status) WARN_UNUSED_RESULT;
